requirements.txt
numpy==1.18.5
scikit-learn==0.23.1
pillow==7.2.0
Flask==1.1.2
requests==2.24.0
opencv-python==4.4.0.46
matplotlib==3.3.3
gcloud==0.18.3
google-api-python-client=1.12.3
flask==1.1.2
jsonpickle==1.4.1
pika==1.1.0
redis==3.5.3
colormath==3.0.0